+++ src/simulation/mesh_loader.cpp
@@ -1,3 +1,8 @@
+// author: @Jan-Piotraschke
+// date: 2023-04-19
+// license: Apache License 2.0
+// version: 0.1.0
+
 #include <iostream>
 #include <string>
 #include <Eigen/Dense>
@@ -42,7 +47,7 @@ Eigen::MatrixXd loadMeshVertices(const std::string& filepath) {
 }
 
 
-Eigen::MatrixXd loadMeshFaces(const std::string& filepath) {
+Eigen::MatrixXi loadMeshFaces(const std::string& filepath) {
     // Create an instance of the Importer class
     Assimp::Importer importer;
 
@@ -53,14 +58,14 @@ Eigen::MatrixXd loadMeshFaces(const std::string& filepath) {
 
     if (!scene) {
         std::cerr << "Failed to load model: " << filepath << std::endl;
-        return Eigen::MatrixXd(0, 0);
+        return Eigen::MatrixXi(0, 0);
     }
 
     // Get the first mesh in the scene
     const aiMesh* mesh = scene->mMeshes[0];
 
         // Create an Eigen matrix to store the face indices
-    Eigen::MatrixXd faces(mesh->mNumFaces, 3);
+    Eigen::MatrixXi faces(mesh->mNumFaces, 3);
 
     // Copy the face indices from the mesh to the Eigen matrix
     for (unsigned int i = 0; i < mesh->mNumFaces; i++) {
